Implemented resizing mechanism.
Implemented resizing mechanism. In order not to shrink the displayed text, there are some minimum limits for the width and height of the application.
Also I have added the window bar back, in order to have access to the maximize/full-screen button. The app title is not on the bar anymore because there is a label with the app name.

Signed-off-by: Cristian Iacob <cristiiacob98@gmail.com>